Two bugs fixed:   - WRITE within the pseudoFS did not return NFS4ERR_ROFS   - Bad XDR encoding in function NFS4_FSALattr_To_Fattr .     This concerned attributes FATTR4_TIME_MODIFY_SET and FATTR4+TIME_ACCESS_SET     that are used only by FSAL_PROXY
Useless checksum has been removed from NFSv4 filehandles.

+++ src/FSAL/FSAL_PROXY/fsal_create.c
@@ -134,7 +134,7 @@ fsal_status_t FSAL_create(fsal_handle_t * parent_directory_handle,      /* IN */
 
   /* Create the owner */
   snprintf(owner_val, FSAL_PROXY_OWNER_LEN, "GANESHA/PROXY: pid=%u ctx=%p file=%llu",
-           getpid(), p_context, p_context->file_counter);
+           getpid(), p_context, (unsigned long long int)p_context->file_counter);
   owner_len = strnlen(owner_val, FSAL_PROXY_OWNER_LEN);
   p_context->file_counter += 1;
